calculator/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
confluent-kafka==2.8.0
grpcio==1.69.0
idna==3.10
protobuf==5.29.3
requests==2.32.3
urllib3==2.3.0
